refactor: replace Replier with Responsive interface and standardize response handling methods across core components for improved clarity and consistency

+++ app/player/internal/app/dev/gate/cmds/user/create_admin_player.go
@@ -68,7 +68,7 @@ func (c *CreateAdminPlayerCommander) Func(ctx core.Context, args map[string]stri
 
 	ctx.Changed()
 
-	_ = ctx.Reply(climod.ModuleID_System, int32(cliseq.SystemSeq_ServerLogout), ctx.UID(), &climsg.SCServerLogout{
+	_ = ctx.Reply(int32(climod.ModuleID_System), int32(cliseq.SystemSeq_ServerLogout), ctx.UID(), &climsg.SCServerLogout{
 		Code: climsg.SCServerLogout_Waiting,
 	})
